// todo-front/src/app/models/todo-app.model.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A universally unique identifier as defined by [RFC 4122](https://tools.ietf.org/html/rfc4122). */
  UUID: any;
  /**
   * A point in time as described by the [ISO
   * 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone.
   */
  Datetime: Date;
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any;
}
// ...
export interface INode {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars["ID"];
}
// ...
export interface ITag extends INode {
  __typename?: "Tag";
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars["ID"];
  id: Scalars["UUID"];
  name: Scalars["String"];
  /** Reads and enables pagination through a set of `TodosTag`. */
  todosTags: ITodosTagsConnection;
}
// ...
export interface ITodosTagsConnection {
  __typename?: "TodosTagsConnection";
  /** A list of `TodosTag` objects. */
  nodes: Array<Maybe<ITodosTag>>;
  /** A list of edges which contains the `TodosTag` and cursor to aid in pagination. */
  edges: Array<ITodosTagsEdge>;
  /** Information to aid in pagination. */
  pageInfo: IPageInfo;
  /** The count of *all* `TodosTag` you could get from the connection. */
  totalCount: Scalars["Int"];
}

export interface ITodosTag extends INode {
  __typename?: "TodosTag";
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars["ID"];
  id: Scalars["UUID"];
  todoId?: Maybe<Scalars["UUID"]>;
  tagId?: Maybe<Scalars["UUID"]>;
  /** Reads a single `Todo` that is related to this `TodosTag`. */
  todo?: Maybe<ITodo>;
  /** Reads a single `Tag` that is related to this `TodosTag`. */
  tag?: Maybe<ITag>;
}

/** A todo */
export interface ITodo extends INode {
  __typename?: "Todo";
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars["ID"];
  id: Scalars["UUID"];
  content: Scalars["String"];
  title: Scalars["String"];
  status?: Maybe<ITodoStatus>;
  createdAt?: Maybe<Scalars["Datetime"]>;
  updatedAt?: Maybe<Scalars["Datetime"]>;
  userId?: Maybe<Scalars["UUID"]>;
  /** Reads a single `User` that is related to this `Todo`. */
  user?: Maybe<IUser>;
  /** Reads and enables pagination through a set of `TodosTag`. */
  todosTags: ITodosTagsConnection;
}

/** A todo */
export interface ITodoTodosTagsArgs {
  first?: Maybe<Scalars["Int"]>;
  last?: Maybe<Scalars["Int"]>;
  offset?: Maybe<Scalars["Int"]>;
  before?: Maybe<Scalars["Cursor"]>;
  after?: Maybe<Scalars["Cursor"]>;
  orderBy?: Maybe<Array<ITodosTagsOrderBy>>;
  condition?: Maybe<ITodosTagCondition>;
  filter?: Maybe<ITodosTagFilter>;
}

export type ITodoStatus = "TO_DO" | "DOING" | "DONE";

export interface IUser extends INode {
  __typename?: "User";
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars["ID"];
  id: Scalars["UUID"];
  name?: Maybe<Scalars["String"]>;
  /** Reads and enables pagination through a set of `Todo`. */
  todos: ITodosConnection;
}

export interface IUserTodosArgs {
  first?: Maybe<Scalars["Int"]>;
  last?: Maybe<Scalars["Int"]>;
  offset?: Maybe<Scalars["Int"]>;
  before?: Maybe<Scalars["Cursor"]>;
  after?: Maybe<Scalars["Cursor"]>;
  orderBy?: Maybe<Array<ITodosOrderBy>>;
  condition?: Maybe<ITodoCondition>;
  filter?: Maybe<ITodoFilter>;
}

/** A connection to a list of `Todo` values. */
export interface ITodosConnection {
  __typename?: "TodosConnection";
  /** A list of `Todo` objects. */
  nodes: Array<Maybe<ITodo>>;
  /** A list of edges which contains the `Todo` and cursor to aid in pagination. */
  edges: Array<ITodosEdge>;
  /** Information to aid in pagination. */
  pageInfo: IPageInfo;
  /** The count of *all* `Todo` you could get from the connection. */
  totalCount: Scalars["Int"];
}

/** A `Todo` edge in the connection. */
export interface ITodosEdge {
  __typename?: "TodosEdge";
  /** A cursor for use in pagination. */
  cursor?: Maybe<Scalars["Cursor"]>;
  /** The `Todo` at the end of the edge. */
  node?: Maybe<ITodo>;
}

/** Information about pagination in a connection. */
export interface IPageInfo {
  __typename?: "PageInfo";
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars["Boolean"];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars["Boolean"];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars["Cursor"]>;
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars["Cursor"]>;
}
// ...
export interface ITodosTagsEdge {
  __typename?: "TodosTagsEdge";
  /** A cursor for use in pagination. */
  cursor?: Maybe<Scalars["Cursor"]>;
  /** The `TodosTag` at the end of the edge. */
  node?: Maybe<ITodosTag>;
}
// ...
export interface ITodoInput {
  id?: Maybe<Scalars["UUID"]>;
  content?: Scalars["String"];
  title?: Scalars["String"];
  status?: Maybe<ITodoStatus>;
  createdAt?: Maybe<Scalars["Datetime"]>;
  updatedAt?: Maybe<Scalars["Datetime"]>;
  userId?: Maybe<Scalars["UUID"]>;
}
// ...
export interface ICusCreateUpdateTodoInput {
  /**
   * An arbitrary string value with no semantic meaning. Will be included in the
   * payload verbatim. May be used to track mutations by the client.
   */
  clientMutationId?: Maybe<Scalars["String"]>;
  iTodo?: Maybe<ITodoInput>;
  tagsId?: Maybe<Array<Maybe<Scalars["UUID"]>>>;
}